+++ src/main/java/domain/piece/kind/PieceStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece.kind;
 
 public enum PieceStatus {
     KING,
