+++ src/main/java/domain/piece/kind/Rook.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.piece.kind;
 
+import domain.piece.Piece;
 import domain.piece.attribute.Color;
 import domain.piece.attribute.point.Direction;
 import domain.piece.attribute.point.Point;
@@ -8,20 +9,20 @@ import java.util.List;
 
 import static domain.piece.attribute.point.Direction.*;
 
-public class Bishop extends Piece {
-    private static final List<Direction> directionList = List.of(DOWN_LEFT, DOWN_RIGHT, UP_LEFT, UP_RIGHT);
+public class Rook extends Piece {
+    private static final List<Direction> rookList = List.of(UP, DOWN, RIGHT, LEFT);
 
-    public Bishop(final Point point, final Color color) {
+    public Rook(final Point point, final Color color) {
         super(point, color);
     }
 
     @Override
     public PieceStatus getStatus() {
-        return PieceStatus.BISHOP;
+        return PieceStatus.ROOK;
     }
 
-    public boolean canMove(Point point) {
+    public boolean canMove(final Point point) {
         final Direction direction = this.point.calculate(point);
-        return directionList.contains(direction);
+        return rookList.contains(direction);
     }
 }
