+++ src/test/java/domain/piece/PieceTest.java
@@ -1,7 +1,7 @@
-package piece;
+package domain.piece;
 
 import domain.piece.attribute.Color;
-import domain.Piece;
+import domain.piece.Piece;
 import fixture.PieceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
