+++ src/test/java/domain/piece/PiecesTest.java
@@ -1,4 +1,8 @@
-import domain.*;
+package domain.piece;
+
+import domain.piece.Piece;
+import domain.piece.Pieces;
+import domain.piece.kind.*;
 import fixture.PieceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
