+++ src/test/java/domain/piece/attribute/point/PointTest.java
@@ -1,3 +1,5 @@
+package domain.piece.attribute.point;
+
 import domain.piece.attribute.point.Index;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
