+++ src/test/java/domain/piece/kind/KingTest.java
@@ -1,6 +1,6 @@
-package piece;
+package domain.piece.kind;
 
-import domain.King;
+import domain.piece.kind.King;
 import domain.piece.attribute.Color;
 import domain.piece.attribute.point.File;
 import domain.piece.attribute.point.Point;
