+++ src/test/java/domain/piece/kind/RookTest.java
@@ -1,6 +1,6 @@
-package piece;
+package domain.piece.kind;
 
-import domain.Rook;
+import domain.piece.kind.Rook;
 import domain.piece.attribute.Color;
 import domain.piece.attribute.point.File;
 import domain.piece.attribute.point.Point;
